#41 Add `Customized` phase of AST node modification
`Customized.$props` stores additional properties initialized during Validation phase. For derivative nodes, these properties are used later to construct their `DerivativeIdentityName`.
It is also expected, that by the time a derivative node is identified, its properties are already `Identified`, so they can be used for identity name construction.

`Customized` had to be added to circumvent an issue on the previous commit, when AST tree gets loops on custom properties assignment, which led to Stackoverflow exceptions